Round out handlers, commands and attributes for SCHEDULE_ENTRY_LOCK commands.
Included handler and commands that aren't supported by the Yale, but
with generic checks to make sure that unsupported commands result in
warnings.

Time and time zone are set automatically. I still need to gauge how the
times in schedules are evaluated (local or UTC) and whether the lock is
aware of DST change dates (the ST command layer does not ask for those
timezone properties as set parameters).